Refactor: unify admin CRM function
